fix: send the control of redirection to the dropdown component
